Update main.py
Updated the code again. For reasons I don't understand, the live game screen wasn't working. So I reverted to an earlier version of the code. BUT there is a catch, you must adjust the HOME score X-Position in the live game section of the code AFTER you set your scale at the top.